refactor(Lotto) : 역할에 맞는 디렉토리로 파일 구분
